Move projectSelectionList android entities to the list package
